requirements.txt
python-dotenv
requests
xmltodict